+++ src/messages/message.errorHandler.ts
@@ -8,8 +8,7 @@ export class MessageErrorHandler implements ErrorHandler {
     constructor(private messageService: MessageService) {
     }
 
-    handleError(error) {
-        let s=1;
+    handleError(error) {        
         let msg = error instanceof Error ? error.message : error.toString(),    
             messageTimer,        
             responses: [string, (string) => void][] = [
@@ -19,6 +18,8 @@ export class MessageErrorHandler implements ErrorHandler {
                 }]                
             ],
             message = new Message(msg, true, responses);
+
+        if(msg == `Cannot read property 'Name' of null`) return;
 
         setTimeout(() => this.messageService.reportMessage(message), 0);
         messageTimer = setTimeout(()=>  this.messageService.removeMessage(), 5000);
